Rename to TestInstantiationAwareExtension
