+++ documentation/src/docs/asciidoc/release-notes/release-notes-5.12.0-M1.adoc
@@ -58,10 +58,11 @@ JUnit repository on GitHub.
   extensions.
 * Allow determining "shared resources" at runtime via the new `@ResourceLock#providers`
   attribute that accepts implementations of `ResourceLocksProvider`.
-* Extensions that participate in the construction of test class instances may override the
-  `getExtensionContextScopeDuringTestClassInstanceConstruction()` method to enable using a
-  test-scoped `ExtensionContext` in `Extension` methods called during test class instance
-  construction. This behavior will become the default in future versions of JUnit.
+* Extensions that implement `TestInstancePreConstructCallback`, `TestInstanceFactory`,
+  `TestInstancePostProcessor`, `ParameterResolver`, or `InvocationInterceptor` may
+  override the `getTestInstantiationExtensionContextScope()` method to enable receiving
+  a test-scoped `ExtensionContext` in `Extension` methods called during test class
+  instantiation. This behavior will become the default in future versions of JUnit.
 * `@TempDir` is now supported on test class constructors.
