+++ junit-jupiter-api/src/main/java/org/junit/jupiter/api/extension/TestInstanceFactory.java
@@ -52,13 +52,13 @@ import org.junit.jupiter.api.extension.ExtensionContext.Store.CloseableResource;
  */
 @FunctionalInterface
 @API(status = STABLE, since = "5.7")
-public interface TestInstanceFactory extends TestClassInstanceConstructionParticipatingExtension {
+public interface TestInstanceFactory extends TestInstantiationAwareExtension {
 
 	/**
 	 * Callback for creating a test instance for the supplied context.
 	 *
 	 * <p>Extensions may override
-	 * {@link #getExtensionContextScopeDuringTestClassInstanceConstruction} for
+	 * {@link #getTestInstantiationExtensionContextScope} for
 	 * revised handling of {@link CloseableResource CloseableResource} and to
 	 * make test-specific data available to your implementation.
 	 *
