+++ junit-jupiter-api/src/main/java/org/junit/jupiter/api/extension/TestInstancePostProcessor.java
@@ -41,13 +41,13 @@ import org.junit.jupiter.api.extension.ExtensionContext.Store.CloseableResource;
  */
 @FunctionalInterface
 @API(status = STABLE, since = "5.0")
-public interface TestInstancePostProcessor extends TestClassInstanceConstructionParticipatingExtension {
+public interface TestInstancePostProcessor extends TestInstantiationAwareExtension {
 
 	/**
 	 * Callback for post-processing the supplied test instance.
 	 *
 	 * <p>Extensions may override
-	 * {@link #getExtensionContextScopeDuringTestClassInstanceConstruction} for
+	 * {@link #getTestInstantiationExtensionContextScope} for
 	 * revised handling of {@link CloseableResource CloseableResource} and to
 	 * make test-specific data available to your implementation.
 	 *
