+++ jupiter-tests/src/test/java/org/junit/jupiter/api/extension/KitchenSinkExtension.java
@@ -65,11 +65,10 @@ public class KitchenSinkExtension implements
 // @formatter:on
 {
 
-	// --- Test Class Instance Construction ------------------------------------
+	// --- Test Class Instantiation --------------------------------------------
 
 	@Override
-	public ExtensionContextScope getExtensionContextScopeDuringTestClassInstanceConstruction(
-			ExtensionContext rootContext) {
+	public ExtensionContextScope getTestInstantiationExtensionContextScope(ExtensionContext rootContext) {
 		return ExtensionContextScope.TEST_METHOD;
 	}
