+++ jupiter-tests/src/test/java/org/junit/jupiter/engine/extension/TestInstancePreConstructCallbackTests.java
@@ -11,7 +11,7 @@
 package org.junit.jupiter.engine.extension;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.extension.TestClassInstanceConstructionParticipatingExtension.ExtensionContextScope.TEST_METHOD;
+import static org.junit.jupiter.api.extension.TestInstantiationAwareExtension.ExtensionContextScope.TEST_METHOD;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -558,8 +558,7 @@ class TestInstancePreConstructCallbackTests extends AbstractJupiterTestEngineTes
 		}
 
 		@Override
-		public ExtensionContextScope getExtensionContextScopeDuringTestClassInstanceConstruction(
-				ExtensionContext rootContext) {
+		public ExtensionContextScope getTestInstantiationExtensionContextScope(ExtensionContext rootContext) {
 			return TEST_METHOD;
 		}
 	}
@@ -570,8 +569,7 @@ class TestInstancePreConstructCallbackTests extends AbstractJupiterTestEngineTes
 		}
 
 		@Override
-		public ExtensionContextScope getExtensionContextScopeDuringTestClassInstanceConstruction(
-				ExtensionContext rootContext) {
+		public ExtensionContextScope getTestInstantiationExtensionContextScope(ExtensionContext rootContext) {
 			return TEST_METHOD;
 		}
 	}
@@ -582,8 +580,7 @@ class TestInstancePreConstructCallbackTests extends AbstractJupiterTestEngineTes
 		}
 
 		@Override
-		public ExtensionContextScope getExtensionContextScopeDuringTestClassInstanceConstruction(
-				ExtensionContext rootContext) {
+		public ExtensionContextScope getTestInstantiationExtensionContextScope(ExtensionContext rootContext) {
 			return TEST_METHOD;
 		}
 	}
